+++ spec/legion/extensions/microsoft_teams/runners/teams_spec.rb
@@ -11,7 +11,7 @@ RSpec.describe Legion::Extensions::MicrosoftTeams::Runners::Teams do
   describe '#list_joined_teams' do
     it 'lists teams for the current user' do
       response = instance_double(Faraday::Response, body: { 'value' => [{ 'id' => 't1', 'displayName' => 'Team A' }] })
-      allow(graph_conn).to receive(:get).with('me/joinedTeams').and_return(response)
+      allow(graph_conn).to receive(:get).with('me/joinedTeams', {}).and_return(response)
 
       result = runner.list_joined_teams
       expect(result[:result]['value'].first['displayName']).to eq('Team A')
@@ -31,7 +31,7 @@ RSpec.describe Legion::Extensions::MicrosoftTeams::Runners::Teams do
   describe '#list_team_members' do
     it 'lists members of a team' do
       response = instance_double(Faraday::Response, body: { 'value' => [{ 'displayName' => 'User A' }] })
-      allow(graph_conn).to receive(:get).with('teams/t1/members').and_return(response)
+      allow(graph_conn).to receive(:get).with('teams/t1/members', { '$top' => 100 }).and_return(response)
 
       result = runner.list_team_members(team_id: 't1')
       expect(result[:result]['value']).not_to be_empty